Pin SQLAlchemy to 2.0.44
Why these changes are being introduced:

We were encountering some errors after SQLAlchemy automatically
updated to 2.0.45 that mentioned "pg_catalog" and "pg_collation".
We use duckdb_engine, a SQLAlchemy engine, to interact with our
DuckDB connection.  It looks as though SQLAlchemy 2.0.45, for reasons
unknown at this time, break that engine, at least the table reflection
which we use on  TIMDEXDataset inititialization.

How this addresses that need:
* Pins SQLAlchemy to 2.0.44 until we learn more or a new version
is released.

Side effects of this change:
* SQLAlchemy will be pinned until touched again.

Relevant ticket(s):
* None