Rename timestamp to embeddings_timestamp
Why these changes are being introduced:

A PR discussion pointed out that `timestamp` was the only column in the new embeddings
schema that, for columns unique to this schema, doesn't prefix with `embeddings_`.  There was
some discussion that for any of those columns it's implied they are part of the embeddings
schema so the `embeddings_` prefix is not necessary.  This would hold true in a pure SQL
world with tables and views acting like namespaces.

But at the level of this python library, where we support quite a bit of kwargs filtering
of read methods, that kind of namespacing is awkward.  We would like to be able to filter by
`run_timestamp` or `embeddings_timestamp`, or in the future maybe `fulltext_timestamp`.  Being
verbose with our columns names simplifies this a bit.

How this addresses that need:

Renames the column `timestamp` to `embedding_timestamp` in the embeddings schema.

Side effects of this change:
* The CLI timdex-embeddings will need an update to what columns it writes to

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-143
* https://mitlibraries.atlassian.net/browse/USE-288